add scope_guard for 'CFRelease(empty)' on iOS (#54)
* add scope_guard `CFRelease(empty)` for allocated CFDictionaryRef

* add `static_cast<int>(…)` for downcast warnings

* bump patch version v0.3.6

+++ ogles_gpgpu/common/proc/mesh.cpp
@@ -117,7 +117,7 @@ void MeshShaderProc::filterRenderSetCoords() {
 }
 
 void MeshShaderProc::filterRenderDraw() {
-    glDrawArrays(triangleKind, 0, vertices.size());
+    glDrawArrays(triangleKind, 0, static_cast<int>(vertices.size()));
 }
 
 void MeshShaderProc::setTriangleKind(GLenum kind) {
